rename: accordion 컴포넌트 폴더 임시로 이름 변경
